react-toastify-notification done

+++ middlewares/imgUploader.js
@@ -3,31 +3,46 @@ const fs = require("fs");
 async function imgUploader(req, res, next) {
   try {
     if (!req.files) {
-      return res.status(400).send("No files were uploaded.");
+      return res.json({
+        status: 500,
+        msg: "Sorry! Something went wrong.",
+      });
     }
     let sampleFile = req.files.id_img;
-    let uploadPath = __dirname + "/../uploads/" + sampleFile.name;
+    console.log(sampleFile);
+    if (sampleFile.size <= 500000) {
+      if (
+        sampleFile.mimetype === "image/jpeg" ||
+        sampleFile.mimetype === "image/jpg" ||
+        sampleFile.mimetype === "image/png" ||
+        sampleFile.mimetype === "image/gif"
+      ) {
+        let uploadPath = __dirname + "/../uploads/" + sampleFile.name;
 
-    // sampleFile.mv(uploadPath, function (err) {
-    //   if (err) {
-    //     return res.status(500).send(err.message);
-    //   } else {
-    //     const urlObject = await imgur.uploadFile(uploadPath);
-    //     fs.unlinkSync(uploadPath);
-    //     req.imgurl = urlObject && urlObject.link;
-    //     //console.log(urlObject);
-    //     console.log(req.imgurl);
-    //   }
-    // });
-    await sampleFile.mv(uploadPath);
-    const urlObject = await imgur.uploadFile(uploadPath);
-    fs.unlinkSync(uploadPath);
-    req.imgurl = urlObject.link;
+        await sampleFile.mv(uploadPath);
+        const urlObject = await imgur.uploadFile(uploadPath);
+        console.log(urlObject);
+        fs.unlinkSync(uploadPath);
+        req.imgurl = urlObject.link;
+        next();
+      } else {
+        return res.json({
+          status: 500,
+          msg: "The image is invalid or not supported.Allowed types: png, jpg, jpeg, gif.",
+        });
+      }
+    } else {
+      return res.json({
+        status: 500,
+        msg: "Image size should be not be greater than 500KB.",
+      });
+    }
   } catch (err) {
-    res.status(500).send("Sorry something went wrong!");
+    res.json({
+      status: 500,
+      msg: "Image upload failed.",
+    });
   }
-
-  next();
 }
 
 module.exports = { imgUploader };
